+++ src/HoYo/Internal/Command.hs
@@ -39,51 +39,51 @@ import Data.Time
 data AddOptions = AddOptions {
   addDirectory  :: TFilePath
   , addName     :: Maybe T.Text
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "move" command to be parsed from the command-line.
 newtype MoveOptions = MoveOptions {
   moveSearch :: BookmarkSearchTerm
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "list" command to be parsed from the command-line.
 data ListOptions = ListOptions {
   listFilterName                :: Maybe T.Text
   , listFilterDirectoryInfix    :: Maybe T.Text
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "clear" command to be parsed from the command-line.
 data ClearOptions = ClearOptions {
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "delete" command to be parsed from the command-line.
 newtype DeleteOptions = DeleteOptions {
   deleteSearch :: BookmarkSearchTerm
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "refresh" command to be parsed from the command-line.
 data RefreshOptions = RefreshOptions {
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "config print" command to be parsed from the command-line.
 data ConfigPrintOptions = ConfigPrintOptions {
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "config reset" command to be parsed from the command-line.
 data ConfigResetOptions = ConfigResetOptions {
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "config set" command to be parsed from the command-line.
 data ConfigSetOptions = ConfigSetOptions {
   setKey        :: T.Text
   , setValue    :: T.Text
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "config add-default" command to be parsed from the command-line.
 data ConfigAddDefaultOptions = ConfigAddDefaultOptions {
   addDefaultDir       :: TFilePath
   , addDefaultName    :: Maybe T.Text
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Options for the "config" command to be parsed from the command-line.
 data ConfigCommand =
@@ -91,13 +91,13 @@ data ConfigCommand =
   | Reset ConfigResetOptions
   | Set ConfigSetOptions
   | AddDefaultBookmark ConfigAddDefaultOptions
-  deriving Show
+  deriving (Show, Eq)
 
 -- | Options for the "check" command to be parsed from the command-line.
 data CheckOptions = CheckOptions {
   checkConfig         :: Bool
   , checkBookmarks    :: Bool
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | The core data-type for the hoyo CLI. The 'Command' is parsed from the command-line,
 -- then 'runCommand' dispatches on the type.
@@ -111,7 +111,7 @@ data Command =
   | ConfigCmd ConfigCommand
   | Check CheckOptions
   | DefaultCommand
-  deriving Show
+  deriving (Show, Eq)
 
 -- | Datatype for representing a command-line settings override.
 data MaybeOverride =
@@ -119,7 +119,7 @@ data MaybeOverride =
   | OverrideTrue
   | NoOverride
   | Conflict
-  deriving Show
+  deriving (Show, Eq)
 
 -- | Combine a config flag with a command-line flag, checking for conflicts.
 combOverride :: Bool -> Bool -> MaybeOverride
@@ -134,7 +134,7 @@ data OverrideOptions = OverrideOptions {
   , overrideDisplayCreationTime     :: MaybeOverride
   , overrideEnableClearing          :: MaybeOverride
   , overrideEnableReset             :: MaybeOverride
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Convert a 'MaybeOverride' to a function on 'Bool'.
 overrideFunc :: MaybeOverride -> (Bool -> Bool)
@@ -169,7 +169,7 @@ data GlobalOptions = GlobalOptions {
   globalConfigPath  :: Maybe TFilePath
   , dataPath        :: Maybe TFilePath
   , overrides       :: OverrideOptions
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | The default behaviour is to override nothing.
 defaultOverrideOptions :: OverrideOptions
@@ -184,7 +184,7 @@ defaultGlobalOptions = GlobalOptions Nothing Nothing defaultOverrideOptions
 data Options = Options {
   optCommand    :: Command
   , optGlobals  :: GlobalOptions
-  } deriving Show
+  } deriving (Show, Eq)
 
 -- | Helper function whenever we need to modify the saved bookmarks.
 --