+++ test/HoYo/Test/Gen.hs
@@ -15,13 +15,22 @@ import Test.QuickCheck.Instances ()
 import System.FilePath
 
 letter :: Gen Char
-letter = elements (['a'..'z'] <> ['A'..'Z'] <> ['0'..'9'])
+letter = elements (['a'..'z'] <> ['A'..'Z'])
+
+letterOrDigit :: Gen Char
+letterOrDigit = elements (['a'..'z'] <> ['A'..'Z'] <> ['0'..'9'])
 
 text :: Gen T.Text
-text = T.pack <$> listOf1 letter
+text = T.pack <$> listOf1 letterOrDigit
+
+textNotNumber :: Gen T.Text
+textNotNumber = T.pack . mconcat <$> sequence [listOf letterOrDigit, (: []) <$> letter, listOf letterOrDigit]
+
+maybeText :: Gen (Maybe T.Text)
+maybeText = liftArbitrary text
 
 string :: Gen String
-string = listOf1 letter
+string = listOf1 letterOrDigit
 
 genFilePath :: Gen TFilePath
 genFilePath = T.pack . joinPath . ("/" :) <$> listOf string
@@ -36,7 +45,7 @@ genFilePath = T.pack . joinPath . ("/" :) <$> listOf string
 --
     -- genTimeZone = TimeZone <$> ((* 60) <$> chooseInt (-12, 12))
                            -- <*> arbitrary
-                           -- <*> vectorOf 3 letter
+                           -- <*> vectorOf 3 letterOrDigit
 
 instance Arbitrary Bookmark where
   arbitrary = Bookmark <$> genFilePath
@@ -53,7 +62,7 @@ instance Arbitrary Bookmarks where
 instance Arbitrary BookmarkSearchTerm where
   arbitrary = oneof [
       SearchIndex <$> chooseInt (1, 1000)
-    , SearchName <$> text
+    , SearchName <$> textNotNumber
     ]
 
 genCommandText :: Gen T.Text
@@ -95,3 +104,73 @@ instance Arbitrary ExecResult where
     , pure ShowHelp
     , ReRun <$> genCommandText
     ]
+
+instance Arbitrary AddOptions where
+  arbitrary = AddOptions <$> genFilePath <*> maybeText
+
+instance Arbitrary MoveOptions where
+  arbitrary = MoveOptions <$> arbitrary
+
+instance Arbitrary ListOptions where
+  arbitrary = ListOptions <$> maybeText <*> maybeText
+
+instance Arbitrary ClearOptions where
+  arbitrary = pure ClearOptions
+
+instance Arbitrary DeleteOptions where
+  arbitrary = DeleteOptions <$> arbitrary
+
+instance Arbitrary RefreshOptions where
+  arbitrary = pure RefreshOptions
+
+instance Arbitrary ConfigPrintOptions where
+  arbitrary = pure ConfigPrintOptions
+
+instance Arbitrary ConfigResetOptions where
+  arbitrary = pure ConfigResetOptions
+
+instance Arbitrary ConfigSetOptions where
+  arbitrary = ConfigSetOptions <$> text <*> text
+
+instance Arbitrary ConfigAddDefaultOptions where
+  arbitrary = ConfigAddDefaultOptions <$> genFilePath <*> maybeText
+
+instance Arbitrary ConfigCommand where
+  arbitrary = oneof [ Print <$> arbitrary
+                    , Reset <$> arbitrary
+                    , Set <$> arbitrary
+                    , AddDefaultBookmark <$> arbitrary
+                    ]
+
+instance Arbitrary CheckOptions where
+  arbitrary = uncurry CheckOptions <$> suchThat arbitrary (uncurry (||))
+
+instance Arbitrary Command where
+  arbitrary = oneof [ Add <$> arbitrary
+                    , Move <$> arbitrary
+                    , List <$> arbitrary
+                    , Clear <$> arbitrary
+                    , Delete <$> arbitrary
+                    , Refresh <$> arbitrary
+                    , ConfigCmd <$> arbitrary
+                    , Check <$> arbitrary
+                    , pure DefaultCommand
+                    ]
+
+instance Arbitrary MaybeOverride where
+  arbitrary = elements [OverrideFalse, OverrideTrue, NoOverride, Conflict]
+
+instance Arbitrary OverrideOptions where
+  arbitrary = OverrideOptions <$> arbitrary
+                              <*> arbitrary
+                              <*> arbitrary
+                              <*> arbitrary
+
+instance Arbitrary GlobalOptions where
+  arbitrary = GlobalOptions <$> liftArbitrary genFilePath
+                            <*> liftArbitrary genFilePath
+                            <*> arbitrary
+
+instance Arbitrary Options where
+  arbitrary = Options <$> arbitrary
+                      <*> arbitrary